Life changes and Optimization

+++ DotnetLogo/NParser/Types/Integer.cs
@@ -7,6 +7,6 @@ namespace NParser.Types
    public class Integer:NetLogoObject
     {
         public override object value { get { return val; } set { val = int.Parse(value.ToString()); } }
-        public  Int64 val { get; set; }
+        public  int val { get; set; }
     }
 }
